requirements.txt
dnspython==2.8.0
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
pydicom==3.0.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
